pyproject.toml
[project]
name = "geminihotel-chat"
version = "0.0.1"
description = "Esse Projeto é um estudo de LLM do Geminai para um Case de Hotel usando ia para um chat bot"
authors = [
    {name = "Pedro Henrique Vieira Barreto"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "pymongo",
    "python-dotenv",
    "uvicorn"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
get_db = "api.services.db_service:conectar_banco"
poplution = "api.data.popular_db:popular_db"
start_chat = "api.services.gemini_ai:main"
start_Api = "run_api:main"

[tool.poetry]
packages = [{ include = "api" }]